More safety nets.
If a wallpaper file is found to be missing, it'll perform an early
database update.

If native resolution isn't possible then it'll pick again without
specifying resolution. Not ideal when you're running 4k monitors and a
720p wallpaper pops up, but it's better than shorting 'feh' of image
arguements.

If it can't find any images, it'll regenerate the image db but also
exit in an error state. (for now)

Script could use less dupliate lines. To convert repeated things into functions.