Added -showdb to print out all known paths and resolutions and then exit. (more of a debug thing)
for the WPRecursion conf option, added an ignore to GNU find to ignore indexing all wallpaper subdirectories featuring the word 'disabled' so directory recursion can still be used for downloaded albums and such.

Fixed an invalid shift in the argument case loop where -regendb would skip an argument after it.