Long needed changes added.
Turned some variables relying on newline delimiters into arrays.

Fixed failure mode if trying to find a spanning wallpaper when there are
multiple screens of different resolutions, fails and tries to span a
single wallpaper across all monitors poorly.

Changed wallpaper scanning process to take checksums of images to
compare with the existing database so they may be skipped if already
seen and with the same checksum as the previous scan. (This does not yet
remove old images, so is prone to error if images are removed and
rescan, but not remove them from the database until corrected).

Some sanity checking of the database state will also need to be done and
only then should a full rebuild (drop table) be done to recreate upon
seeing a severe error in expected db formatting or output.