Further sanity checking.
Added a database structure associative array which is looped through in
comparison with the sqlite3 database checking a column exists and with
the correct datatype. If not seen, it gets created. This is done in case
a new version of the script adds a new column or changes a datatype.
Saves rescanning every detail for large wallpaper libraries.